POST is working!!!

+++ migrations/20231205220339_Users.js
@@ -17,6 +17,9 @@ exports.up = function(knex) {
  * @returns { Promise<void> }
  */
 exports.down = function(knex) {
-  return knex.schema.dropTable('Users');
+  return knex.schema.hasTable('Users').then(function(exists) {
+    if (exists) {
+      return knex.schema.dropTable('Users');
+    }
+  });
 };
-
